Feat: Return category in model transaction, added end date to budget

+++ FinanceManager/Models/Budget.cs
@@ -33,7 +33,7 @@ namespace FinanceManager.Models
         }
         public int GetProgressPercent()
         {
-            int Percent = SpentAmount / LimitAmount * 100;
+            int Percent = (int)((double)SpentAmount / LimitAmount * 100);
             return Percent;
         }
